chore: Remove the sign in but keep the identity validation
